[#21] time converter domain 레이어로 옮김
